pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "customer-support-env"
version = "0.1.0"
description = "Customer Support Triage Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
